+++ rollup.config.js
@@ -4,12 +4,14 @@ import resolve from '@rollup/plugin-node-resolve';
 import typescript from '@rollup/plugin-typescript';
 import { terser } from 'rollup-plugin-terser';
 
+const { fflate, ...dependencies } = require('./package.json').dependencies
+
 export default {
   input: join('src', 'index.ts'),
   external: [
     // These Node.js interenals are external to our bundles…
     // …as are the dependencies listed in our package.json.
-    ...Object.keys(require('./package.json').dependencies),
+    ...Object.keys(dependencies),
   ],
   output: [
     { file: join('build', 'index.cjs.js'), format: 'cjs' },
@@ -18,9 +20,6 @@ export default {
   plugins: [
     resolve({
       extensions: ['.ts'],
-      customResolveOptions: {
-        moduleDirectory: 'src',
-      },
       preferBuiltins: true,
     }),
     typescript(),