+++ src/services/payments/appStore.ts
@@ -46,18 +46,32 @@ export default (
 
   async getNotifications(options) {
     return (
-      await client.get(httpAuth, '/appStore/receivedNotifications', options)
+      await client.get(
+        httpAuth,
+        `/appStore/receivedNotifications/${options?.rql || ''}`,
+        options
+      )
     ).data;
   },
 
   async getReceipts(options) {
-    return (await client.get(httpAuth, '/appStore/receivedReceipts', options))
-      .data;
+    return (
+      await client.get(
+        httpAuth,
+        `/appStore/receivedReceipts/${options?.rql || ''}`,
+        options
+      )
+    ).data;
   },
 
   async getSharedSecrets(options) {
-    return (await client.get(httpAuth, '/appStore/sharedSecrets', options))
-      .data;
+    return (
+      await client.get(
+        httpAuth,
+        `/appStore/sharedSecrets/${options?.rql || ''}`,
+        options
+      )
+    ).data;
   },
 
   async createSharedSecret(requestBody, options) {