+++ src/services/payments/types.ts
@@ -1,4 +1,3 @@
-/* eslint-disable camelcase */
 import { FindAllIterator } from '../../services/helpers';
 import { RQLString } from '../../rql';
 import type {
@@ -451,7 +450,7 @@ export interface PaymentsAppStoreService {
    * @returns PagedResult<AppStoreNotification>
    */
   getNotifications(
-    options?: OptionsBase
+    options?: OptionsWithRql
   ): Promise<PagedResult<AppStoreNotification>>;
   /**
    * Get a list of receipts received and verified by the App Store
@@ -464,7 +463,7 @@ export interface PaymentsAppStoreService {
    * `VIEW_APP_STORE_RECEIPTS` | `global` | **Required** for this endpoint
    * @returns PagedResult<AppStoreReceipt>
    */
-  getReceipts(options?: OptionsBase): Promise<PagedResult<AppStoreReceipt>>;
+  getReceipts(options?: OptionsWithRql): Promise<PagedResult<AppStoreReceipt>>;
 
   /**
    * Get a list of shared secrets
@@ -1120,6 +1119,7 @@ export interface PlayStorePurchaseRecord {
   userId: ObjectId;
   id: ObjectId;
   creationTimestamp: Date;
+  updateTimestamp: Date;
 }
 
 export interface PlayStoreDeveloperNotificationSchema {
@@ -1185,20 +1185,18 @@ export interface PlayStoreSubscriptionPurchaseRecordSchema {
   purchaseInfo: PlayStoreSubscriptionPurchaseSchema;
   id: string;
   creationTimestamp: Date;
+  updateTimestamp: Date;
+}
+
+export interface PlayStoreSubscriptionReceiptRecordSchema {
+  receipt: PlayStoreReceiptSchema;
+  userId: ObjectId;
+  id: string;
+  creationTimestamp: Date;
+  updateTimestamp: Date;
 }
 
 export interface PaymentsPlayStoreHistoryService {
-  /**
-   * Get a list of recorded Play Store purchase requests
-   *
-   * Permission | Scope | Effect
-   * - | - | -
-   * `VIEW_PLAY_STORE_PURCHASES` | `global` | **Required** for this endpoint
-   * @returns PagedResult<PlayStorePurchaseRecord>
-   */
-  purchases(
-    options?: OptionsWithRql
-  ): Promise<PagedResult<PlayStorePurchaseRecord>>;
   /**
    * Get a list of recorded Play Store developer notifications received
    *
@@ -1221,6 +1219,17 @@ export interface PaymentsPlayStoreHistoryService {
   purchaseInfos(
     options?: OptionsWithRql
   ): Promise<PagedResult<PlayStoreSubscriptionPurchaseRecordSchema>>;
+  /**
+   * Get a list of the received Play Store purchase receipts
+   *
+   * Permission | Scope | Effect
+   * - | - | -
+   * `VIEW_PLAY_STORE_PURCHASE_RECEIPTS` | `global` | **Required** for this endpoint
+   * @returns PagedResult<PlayStorePurchaseRecord>
+   */
+  purchaseReceipts(
+    options?: OptionsWithRql
+  ): Promise<PagedResult<PlayStorePurchaseRecord>>;
 }
 
 export interface PlayStoreSubscription {